set admins and restaurants users authentication
