+++ src/main/java/com/itep/restaurant_service/repositories/RestaurantRepository.java
@@ -1,8 +1,11 @@
 package com.itep.restaurant_service.repositories;
 
+import com.itep.restaurant_service.repositories.entities.AdminEntity;
 import com.itep.restaurant_service.repositories.entities.RestaurantEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
 
 @Repository
 public interface RestaurantRepository extends JpaRepository<RestaurantEntity, Long> {
